Cleaning up the displaying of the index page of the different categories some more.
Added a "Kind" model which will hold the different 'types' that are within the DRM free and ebook items, e.g. .mobi, .cbz, .epub.

I think "Kind" is a bad name for this model but the polymorphic relationship enabled me to rename it to 'platforms'. Perhaps some
refactoring later can fix this name, though.